+++ build.gradle.kts
@@ -1,86 +1,58 @@
-/*
- * This file was generated by the Gradle 'init' task.
- */
+import net.civmc.civgradle.common.util.civRepo
 
 plugins {
     `java-library`
     `maven-publish`
-    id("io.papermc.paperweight.userdev") version "1.3.1"
+    id("net.civmc.civgradle.plugin") version "1.0.0-SNAPSHOT"
 }
 
-group = "net.civmc"
-version = "2.0.0-SNAPSHOT"
-description = "CivChat2"
-
-repositories {
-	fun civRepo(name: String) {
-    	maven {
-    		url = uri("https://maven.pkg.github.com/CivMC/${name}")
-    		credentials {
-    			// These need to be set in the user environment variables
-    			username = System.getenv("GITHUB_ACTOR")
-    			password = System.getenv("GITHUB_TOKEN")
-    		}
-    	}
-    }
-    mavenCentral()
-    civRepo("CivModCore")
-    civRepo("NameLayer")
-
-    maven("https://papermc.io/repo/repository/maven-public/")
-    maven("https://repo.maven.apache.org/maven2/")
+// Temporary hack:
+// Remove the root build directory
+gradle.buildFinished {
+	project.buildDir.deleteRecursively()
 }
 
-dependencies {
-	paperDevBundle("1.18-R0.1-SNAPSHOT")
-
-	implementation("net.civmc:civmodcore:2.0.0-SNAPSHOT:dev-all")
-    implementation("net.civmc:namelayer-spigot:3.0.0-SNAPSHOT:dev")
-    implementation("org.apache.logging.log4j:log4j-core:2.16.0")
+allprojects {
+	group = "net.civmc.civchat2"
+	version = "2.0.0-SNAPSHOT"
+	description = "CivChat2"
 }
 
-java {
-	toolchain.languageVersion.set(JavaLanguageVersion.of(17))
-}
+subprojects {
+	apply(plugin = "net.civmc.civgradle.plugin")
+	apply(plugin = "java-library")
+	apply(plugin = "maven-publish")
 
-tasks {
-	build {
-		dependsOn(reobfJar)
-	}
-
-	compileJava {
-		options.encoding = Charsets.UTF_8.name() // We want UTF-8 for everything
-
-		// Set the release flag. This configures what version bytecode the compiler will emit, as well as what JDK APIs are usable.
-		// See https://openjdk.java.net/jeps/247 for more information.
-		options.release.set(17)
-	}
-	javadoc {
-		options.encoding = Charsets.UTF_8.name() // We want UTF-8 for everything
-	}
-	processResources {
-		filteringCharset = Charsets.UTF_8.name() // We want UTF-8 for everything
-	}
-
-	test {
-		useJUnitPlatform()
-	}
-}
-
-publishing {
-	repositories {
-		maven {
-			name = "GitHubPackages"
-			url = uri("https://maven.pkg.github.com/CivMC/CivChat2")
-			credentials {
-				username = System.getenv("GITHUB_ACTOR")
-				password = System.getenv("GITHUB_TOKEN")
-			}
+	java {
+		toolchain {
+			languageVersion.set(JavaLanguageVersion.of(17))
 		}
 	}
-	publications {
-		register<MavenPublication>("gpr") {
-			from(components["java"])
+
+	repositories {
+		mavenCentral()
+        civRepo("CivMC/CivModCore")
+        civRepo("CivMC/NameLayer")
+
+        maven("https://papermc.io/repo/repository/maven-public/")
+        maven("https://repo.maven.apache.org/maven2/")
+	}
+
+	publishing {
+		repositories {
+			maven {
+				name = "GitHubPackages"
+				url = uri("https://maven.pkg.github.com/CivMC/CivChat2")
+				credentials {
+					username = System.getenv("GITHUB_ACTOR")
+					password = System.getenv("GITHUB_TOKEN")
+				}
+			}
+		}
+		publications {
+			register<MavenPublication>("gpr") {
+				from(components["java"])
+			}
 		}
 	}
 }
