+++ paper/src/main/java/vg/civcraft/mc/civchat2/ChatStrings.java
