+++ paper/src/main/java/vg/civcraft/mc/civchat2/commands/Afk.java
