+++ paper/src/main/java/vg/civcraft/mc/civchat2/utility/CivChat2FileLogger.java
