+++ paper/src/main/java/vg/civcraft/mc/civchat2/utility/ScoreboardHUD.java
