+++ paper/src/main/resources/log4j2.xml
@@ -1,0 +1,73 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<Configuration status="WARN" packages="com.mojang.util">
+
+    <Appenders>
+
+        <Console name="WINDOWS_COMPAT" target="SYSTEM_OUT"></Console>
+
+        <Queue name="TerminalConsole">
+
+            <PatternLayout pattern="[%d{HH:mm:ss} %level]: %msg%n" />
+
+        </Queue>
+
+        <RollingRandomAccessFile name="File" fileName="logs/latest.log" filePattern="logs/%d{yyyy-MM-dd}-%i.log.gz">
+
+            <PatternLayout pattern="[%d{HH:mm:ss}] [%t/%level]: %msg%n" />
+
+            <Policies>
+
+                <TimeBasedTriggeringPolicy />
+
+                <OnStartupTriggeringPolicy />
+
+            </Policies>
+
+        </RollingRandomAccessFile>
+
+		        <RollingRandomAccessFile name="ChatLogsFile" fileName="logs/ChatLogs/latest.log" filePattern="logs/ChatLogs/%d{yyyy-MM-dd}-%i.log.gz">
+
+            <PatternLayout pattern="[%d{HH:mm:ss}] [Chat]: %msg%n" />
+
+            <Policies>
+
+                <TimeBasedTriggeringPolicy />
+
+                <OnStartupTriggeringPolicy />
+
+            </Policies>
+
+        </RollingRandomAccessFile>
+
+    </Appenders>
+
+    <Loggers>
+
+	    <Logger name="ChatLogs" level="info" additivity="false">
+
+            <AppenderRef ref="ChatLogsFile"/>
+
+            <AppenderRef ref="TerminalConsole" level="info"/>
+
+        </Logger>
+
+        <Root level="info">
+
+            <filters>
+
+                <MarkerFilter marker="NETWORK_PACKETS" onMatch="DENY" onMismatch="NEUTRAL" />
+
+            </filters>
+
+            <AppenderRef ref="WINDOWS_COMPAT" level="info"/>
+
+            <AppenderRef ref="File"/>
+
+            <AppenderRef ref="TerminalConsole" level="info"/>
+
+        </Root>
+
+    </Loggers>
+
+</Configuration>
