+++ paper/src/main/resources/plugin.yml
@@ -1,6 +1,6 @@
-name: CivChat2
+name: ${name}
 main: vg.civcraft.mc.civchat2.CivChat2
-version: 2.0.0-SNAPSHOT
+version: ${version}
 api-version: 1.16
 authors:
 - jjj5311
